Infraestructura/docker/python/odbc/sql_server/requirements.txt
aioodbc==0.5.0
python-dotenv==1.0.1